#1437 refactor(agent): colocate sandbox lease service
